requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.12.1
pillow>=10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
numpy>=1.26.0,<2.0
opencv-python-headless>=4.9.0
redis==5.0.1
